pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flex_desal_analysis"
readme = "README.md"
version = "0.1.0"
description = "Python package for conducting analysis for NAWI 3.25"
requires-python = ">=3.11,<3.12"
dependencies = [
    "idaes-pse>=2.10",
    "pyomo>=6.9.3",
    "pint>=0.24",
    "pyyaml",
    "pandas",
    "numpy",
    "matplotlib",
    "scipy"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]